+++ src/main/kotlin/de/qualersoft/jmeter/gradleplugin/task/JMeterSetupTask.kt
@@ -11,9 +11,9 @@ import org.gradle.api.artifacts.ResolvedArtifact
 import org.gradle.api.artifacts.ResolvedDependency
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.tasks.CacheableTask
-import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.Nested
 import org.gradle.api.tasks.OutputDirectory
+import org.gradle.api.tasks.OutputFile
 import org.gradle.api.tasks.TaskAction
 import java.io.File
 import java.util.jar.JarFile
@@ -32,8 +32,12 @@ open class JMeterSetupTask : DefaultTask() {
   private val jmExtDir = jmLibDir.resolve("ext")
   private val jmJUnitDir = jmLibDir.resolve("junit")
 
-  @get:Internal
-  internal val jmJar: RegularFileProperty = project.objects.fileProperty()
+  private val sourceJmJar by lazy { getJMeterLib() }
+
+  @get:OutputFile
+  internal val jmJar: RegularFileProperty = project.objects.fileProperty().value {
+    jmBinDir.resolve("${jmTool.name}-${jmTool.version}.jar")
+  }
 
   init {
     group = "jmeter"
@@ -44,7 +48,7 @@ open class JMeterSetupTask : DefaultTask() {
     prepareDirectories()
 
     // copy jmeter-runner to bin dir
-    jmJar.set(getJMeterLib().copyToDir(jmBinDir))
+    sourceJmJar.copyTo(jmJar.asFile.get(), true)
 
     val resourceJar = getJMeterResourceLib()
     CopyResource.extractJarToDir(JarFile(resourceJar), jmToolDir)
@@ -59,7 +63,8 @@ open class JMeterSetupTask : DefaultTask() {
   private fun getJMeterLib(): File {
     val artifacts: Set<ResolvedArtifact> = project.configurations
       .getByName(JMETER_RUNNER)
-      .resolvedConfiguration.resolvedArtifacts
+      .resolvedConfiguration
+      .resolvedArtifacts
 
     return findArtifactMatch(artifacts, jmTool.group, jmTool.name)
   }
@@ -70,7 +75,8 @@ open class JMeterSetupTask : DefaultTask() {
   private fun getJMeterResourceLib(): File {
     val artifacts: Set<ResolvedArtifact> = project.configurations
       .getByName(JMETER_RUNNER)
-      .resolvedConfiguration.resolvedArtifacts
+      .resolvedConfiguration
+      .resolvedArtifacts
 
     val toolConfNot = jmTool.createToolConfigDependencyNotion()
     val toolConfName = toolConfNot["name"]!!
@@ -91,26 +97,22 @@ open class JMeterSetupTask : DefaultTask() {
   private fun resolveAndCopyExtensionLibs() {
     val resolvedExtensions = mutableListOf<ResolvedDependency>()
     project.configurations.getByName(JMETER_PLUGIN_DEPENDENCY)
-      .resolvedConfiguration.firstLevelModuleDependencies.flatMap {
+      .resolvedConfiguration
+      .firstLevelModuleDependencies
+      .flatMap {
         resolvedExtensions.add(it)
         it.moduleArtifacts
-      }.map {
-        it.file
-      }.forEach {
-        it.copyToDir(jmExtDir)
       }
-    resolvedExtensions.flatMap {
-      it.children
-    }.filterNot {
+      .map { it.file }
+      .forEach { it.copyToDir(jmExtDir) }
+
+    resolvedExtensions
+      .flatMap { it.children }
       // only take dependencies that were not already copied earlier
-      resolvedExtensions.contains(it)
-    }.flatMap {
-      it.allModuleArtifacts
-    }.map {
-      it.file
-    }.forEach {
-      it.copyToDir(jmLibDir)
-    }
+      .filterNot { resolvedExtensions.contains(it) }
+      .flatMap { it.allModuleArtifacts }
+      .map { it.file }
+      .forEach { it.copyToDir(jmLibDir) }
   }
 
   /**
@@ -118,11 +120,10 @@ open class JMeterSetupTask : DefaultTask() {
    */
   private fun resolveAndCopyToolLibs() {
     project.configurations.getByName(JMETER_LIB_DEPENDENCY)
-      .resolvedConfiguration.resolvedArtifacts.map {
-        it.file
-      }.forEach {
-        it.copyToDir(jmLibDir)
-      }
+      .resolvedConfiguration
+      .resolvedArtifacts
+      .map { it.file }
+      .forEach { it.copyToDir(jmLibDir) }
   }
 
   /**